The type of setups used is now written to the restart file.
Introduced gd.zeros and gd.empty methods.

Allocations of Numeric arrays has been switched from num.zeros to
num.empty where this makes sence.  This has three purposes:

1) It is much faster (especially for large arrays).
2) I think it makes the code easier to read when when we are explicit
about whether an array needs to be initialized to zero or not.
3) It helps in finding some types of bugs.

In debug mode, num.empty is replaced by a function that returns arrays
containing the value 117.

Restarting from file is now much faster.  Allocation of some big
arrays and the calculation of localized orbitals is delayed to the
point where a calculation (after a change of the atomic positions)
starts.  I had to move some code to some strange places - I'm not
proud of it!  I'll clean it up later.



git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@508 6419b465-d902-0410-997c-be2266baf723