Hello! This is my first check in, so I hope that I won't crash anything. Here is the all electron KLI-functional that I have implemented. Feel free to comment and make suggestions.
Few things that I should mention:
1. I introduced xc.orbital_dependent flag, that is used for KLI-functional. This
creates an alternative structure to xc.hybrid, so there might be a place
to make code more uniform

2.
The KLI-potential cannot calculated every step, since that kind of calculation
didn't converge.

3.
The code is still in developing stage, but I wanted to make this commit to
see if I am going in to the right direction. 
 


git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@567 6419b465-d902-0410-997c-be2266baf723