GLLB will now use orbital_dependent as it should be.
I cleaned up, structured and commented the GLLB code, this makes it easier to add new options later (for 
example orbital dependent correlation functionals).

This is by no means the final version, the XCCorrections part still needs to be cleaned up and the GLLB 
functional needs to be tested.

I will perform a group of tests to test this new functional, I will discuss the results later when I have 
them. The atomization energies would be a good benchmark, but the spin-polarized calculation is still 
missing. I recall that there was discussion about converting the 1D-code to spin-polarized, is that going 
on? I think it would be rather easy thing as far as all_electron.py is consired, but hard when one needs to 
consider generator.py also.



git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@1199 6419b465-d902-0410-997c-be2266baf723