Hello!
Lately, I have been commenting and testing the gllb-code. The work is not yet done, but i realized that I have to 
keep up with the current version to make things easier in the future. 

Following things have changed:

-For convience, I removed the kli.py from the main directory, and created new gllb directory and split
the kli.py to several logical files. I have tried to write to code as general as possible, so in principle
the code should be easy to expand.

-The commenting and testing is not done, so do not rely on the results.

-For some reason, the convergence is rather poor at first. This needs to be investigated, maybe plotting the 
3D-potentials and densities. This might have something to do with the fact that I am doing linear mixing
for Vxc-potential, while the code is doing pulay mixing for density (and density matrices). I know a way to tests
wheter this is true.

-The parameter list from read_setup is getting larger. I generalized the setup-part for gllb in such way that
any argument from xml file starting with GLLB_ will be stored to one dictionary, which will be passed
on to XCNonLocalCorrections.

-I use factory design pattern to get the correct type of "non-local"-functional. 

-I expanded the Slater potential calculation in KLI using legendres polynomials, so Function1D.py class in not 
needed. It is still needed in some parts of my experimental codes so I will not remove it at least now.

-I removed the non-working 3D-KLI at least for now.

Feel free to give comments and critique about the implementation.

Best Regards,
Mikael Kuisma



git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@1251 6419b465-d902-0410-997c-be2266baf723