Warning on the use of a.conj() versus npy.conj(a)
If a is real, a.conj() returns a pointer to the original array.
If this is *not* desired, use npy.conj(a) which will always make a copy.

The problem arises if you try to do something like a += a.T.conj(), which will not produce an error, but will give the wrong result, as the thing you add is a view into the array you a modifying.

Nasty


git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@1823 6419b465-d902-0410-997c-be2266baf723