Bug fix: the test suite would occationally hang when running in parallel.  The previous fix did not fully solve the problem.
When a parallel test sequentially writes and reads a file, and REUSE
THE SAME FILE NAME, a barrier call must be made not just after each
write, but also before the next write.  Otherwise the master process
may overwrite the file with the next test, before one of the other
tasks has completed reading it.

Closes #87.



git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@3608 6419b465-d902-0410-997c-be2266baf723